source/requirements.txt
flask
flask-cors
torch
transformers
bitsandbytes
accelerate
opencv-python-headless
pillow
numpy
apscheduler
gunicorn